+++ src/mlpack/methods/cf/decomposition_policies/decomposition_policies.hpp
@@ -20,5 +20,7 @@
 #include "svd_complete_method.hpp"
 #include "svd_incomplete_method.hpp"
 #include "svdplusplus_method.hpp"
+#include "quic_svd_method.hpp"
+#include "block_krylov_svd_method.hpp"
 
 #endif