+++ HISTORY.md
@@ -3,7 +3,9 @@
 
   * Fix PReLU ann layer (#3420).
 
-  * Update outdated code in tutorials (#3398).
+  * Add CF decomposition methods: `QUIC_SVDPolicy` and `BlockKrylovSVDPolicy` (#3413, #3404).
+
+  * Update outdated code in tutorials (#3398, #3401).
 
   * Bugfix for non-square convolution kernels (#3376).